+++ test/index.test.ts
@@ -1,23 +1,17 @@
-/* eslint-disable jest/expect-expect */
+import { fileURLToPath } from 'url';
+import { dirname } from 'path';
 import path from 'path';
 
-import {
-	Array,
-	Boolean,
-	InstanceOf,
-	Literal,
-	Number,
-	Partial,
-	Record,
-	String,
-	Union,
-} from 'runtypes';
+import { z } from 'zod';
 
-import { EnumRecord } from '../src/types';
-import { DEFAULT_PREFIX } from '../src/env';
-import load from '../src/load-config';
+import { EnumRecord } from '../src/types.js';
+import { DEFAULT_PREFIX } from '../src/env.js';
+import load from '../src/load-config.js';
 
-import spec from './fixtures/config/spec';
+import spec from './fixtures/config/spec.js';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
 // for testing
 class Foo {}
@@ -136,13 +130,13 @@ describe('getConfig', () => {
 	it('should coerce environment value to number, boolean, and date', () => {
 		process.env.CONFIG_port = '1000';
 		process.env.CONFIG_debug = 'true';
-		process.env.CONFIG_startTime = '12-11-2021T00:00:00';
+		process.env.CONFIG_startTime = '2021-12-11T00:00:00';
 		const result = load({
 			path: '',
-			schema: Record({
-				port: Number,
-				debug: Boolean,
-				startTime: InstanceOf(Date),
+			schema: z.object({
+				port: z.number(),
+				debug: z.boolean(),
+				startTime: z.date(),
 			}),
 		});
 		expect(result.port).toBe(1000);
@@ -175,38 +169,38 @@ describe('getConfig', () => {
 		);
 	});
 
-    it('should use default values for env templates', () => {
-        process.env.URL_CA = 'test'
-        const result = load({
-            path: path.join(__dirname, 'fixtures', 'config', 'with-env'),
-            sources: ['default', 'env-defaults'],
-            schema: spec
-        })
-        expect(result).toMatchObject(
-            expect.objectContaining<typeof result>({
-                database: {
-                    host: 'database.server',
-                    port: 4000,
-                },
-                sites: {
-                  CA: {
-                      url: 'test'
-                  },
-                  US: {
-                      url: 'https://us.com'
-                  }
-                }
-            })
-        );
-    })
+	it('should use default values for env templates', () => {
+		process.env.URL_CA = 'test';
+		const result = load({
+			path: path.join(__dirname, 'fixtures', 'config', 'with-env'),
+			sources: ['default', 'env-defaults'],
+			schema: spec,
+		});
+		expect(result).toMatchObject(
+			expect.objectContaining<typeof result>({
+				database: {
+					host: 'database.server',
+					port: 4000,
+				},
+				sites: {
+					CA: {
+						url: 'test',
+					},
+					US: {
+						url: 'https://us.com',
+					},
+				},
+			})
+		);
+	});
 
 	it('should not error on invalid env coercion', () => {
 		process.env.CONFIG_foo = 'true';
 		expect(() =>
 			load({
 				path: '',
-				schema: Record({
-					foo: InstanceOf(Foo),
+				schema: z.object({
+					foo: z.instanceof(Foo),
 				}),
 			})
 		).toThrow();
@@ -217,9 +211,9 @@ describe('getConfig', () => {
 		process.env.CONFIG_bar = 'env-bar';
 		const result = load({
 			path: '',
-			schema: Record({
-				foo: Union(Literal('foo'), Literal('baz')),
-				bar: Literal('bar'),
+			schema: z.object({
+				foo: z.union([z.literal('foo'), z.literal('baz')]),
+				bar: z.literal('bar'),
 			}),
 			defaults: {
 				foo: 'baz',
@@ -237,13 +231,17 @@ describe('getConfig', () => {
 		process.env.CONFIG_bar = 'bar';
 		const result = load({
 			path: '',
-			schema: Record({
-				foo: String,
-			}).And(
-				Partial({
-					bar: String,
+			schema: z
+				.object({
+					foo: z.string(),
 				})
-			),
+				.merge(
+					z
+						.object({
+							bar: z.string(),
+						})
+						.partial()
+				),
 		});
 		expect(result).toMatchObject({
 			foo: 'foo',
@@ -257,9 +255,9 @@ describe('getConfig', () => {
 		const result = load({
 			path: '',
 			schema: EnumRecord(
-				Union(Literal('US'), Literal('CA')),
-				Record({
-					foo: String,
+				z.enum(['US', 'CA']),
+				z.object({
+					foo: z.string(),
 				})
 			),
 		});
@@ -382,8 +380,8 @@ describe('getConfig', () => {
 
 		const result = load({
 			path: '',
-			schema: Record({
-				nameMatch: InstanceOf(RegExp),
+			schema: z.object({
+				nameMatch: z.instanceof(RegExp),
 			}),
 			defaults: {
 				nameMatch: /asdf/,
@@ -401,8 +399,8 @@ describe('getConfig', () => {
 			path: '',
 			format: 'json',
 			sources: ['ENV'],
-			schema: Record({
-				arrayValue: Array(String),
+			schema: z.object({
+				arrayValue: z.array(z.string()),
 			}),
 		});
 		expect(result).toMatchObject<typeof result>(